requirements.txt
pyyaml
pip
